Patched s3Client.destroy() in list-object-versions.js

+++ nodes/list-object-versions/list-object-versions.js
@@ -120,7 +120,7 @@ function inputHandler(n, RED) {
       msg.bucket = payloadConfig.Bucket;
       send(msg);
     } finally {
-      s3Client.destroy();
+      if(s3Client) s3Client.destroy();
       /* Dereference vars */
       s3Client = null;
       /*********************/
